Record F6/Output dense IP Vitis HLS C-sim/C-synth results
F6 (PE_COUNT=4 partial-sum): II=4, DSP=11, FF=1673, LUT=1988,
Fmax=111.66MHz -- confirms the accumulation-limited II=4 floor holds
independent of N_IN, same fadd-latency mechanism as C5.

Output (MAC stage + separate softmax stage): MAC stage confirmed at
II=4 (same as C5/F6); softmax's own scalar sum_exp accumulation lands
on the same II=5 accumulator-latency pattern as C3/C5's un-split
baselines, left unsplit since PE_COUNT=4 over a 10-term reduction
isn't worth the resource cost. Top-level latency 999/interval 1000
(4 sequential stages, not pipelined together, by design).

Also fixes a real build error: std::expf isn't in Vitis HLS's bundled
MinGW <cmath> -- switched to std::exp (valid float overload).

Ignore the two new generated Vitis HLS project directories
(dense_f6_proj, dense_output_proj), same pattern as C5/C1.

+++ hls/dense_output/src/dense_output.cpp
@@ -63,7 +63,7 @@ void dense_output(
     float exp_vals[N_OUT];
     float sum_exp = 0;
     for (int j = 0; j < N_OUT; j++) {
-        exp_vals[j] = std::expf(logits[j] - max_logit);
+        exp_vals[j] = std::exp(logits[j] - max_logit);
         sum_exp += exp_vals[j];
     }
